% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gen_func}
\alias{gen_func}
\title{Simulate example data for Additive Factor Model}
\usage{
gen_func(theta)
}
\arguments{
\item{theta}{vector of coefficients for the basis functions. Length of the vector should be even.}
}
\description{
Generate a function from linear combinations of trigonometric functions.
The functions are centered, i.e. E(g(Z)) = 0, if Z is normally distributed.
}
\examples{
g <- gen_func(rnorm(8))
z <- seq(-3, 3, l = 100)
plot(g(z) ~ z, type = "l")
}
