% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{plot.sql}
\alias{plot.sql}
\title{Plot for Statistical Quantile Learning}
\usage{
\method{plot}{sql}(object, ...)
}
\arguments{
\item{object}{a fitted sql object produced by SQL().}

\item{...}{other graphics parameters to pass on to the plotting commands.
If this is not provided, predictions are perform over a grid of m=n normal quantiles.}
}
\description{
Takes a fitted sql object produced by SQL() and plot the factors.
}
