% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{predict.sql}
\alias{predict.sql}
\title{Predictions for Statistical Quantile Learning}
\usage{
\method{predict}{sql}(object, new_data, type = c("terms", "response"), scaling = pnorm)
}
\arguments{
\item{object}{a fitted sql object produced by SQL().}

\item{new_data}{a matrix, with q columns, containing the values (of the latent variables) at which predictions are required.}

\item{type}{type of predictions, with choices "terms" or "response".}

\item{scaling}{A probability distribution function corresponding to the distribution of new_data.}
}
\description{
Obtain predictions from a fitted sql object produced by SQL().
}
