% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{print.sql}
\alias{print.sql}
\title{Print a Statistical Quantile Learning}
\usage{
print.sql(object)
}
\arguments{
\item{object}{a fitted sql object produced by SQL().}
}
\description{
The default print method for a sql object
}
