% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{simulate_afm}
\alias{simulate_afm}
\title{Simulate example data for Additive Factor Model}
\usage{
simulate_afm(n, p, q = 1, sde = 1, factor = NULL, generator = NULL)
}
\arguments{
\item{n}{sample size.}

\item{p}{number of variables.}

\item{q}{number of factors.}

\item{sde}{standard deviation of the errors.}

\item{factor}{a n by q matrix of latent factors.
If not specified, the factors are generated as independent normally distributed variables.}

\item{generator}{a list of length q functions.
Each function takes as input a vector of length m and produce a m by p matrix.
If not specified, the factors are generated as using random trigonometric functions.}
}
\description{
Function used to simulated data sets to illustrate the use of SQL.
The generating function are simulated randomly using trigonometric basis functions.
The factor are simulated as independent normally distributed variables (unless specified otherwise).
}
