% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CreateTable.R
\docType{methods}
\name{sqlCreateTable}
\alias{sqlCreateTable}
\alias{sqlCreateTable,DBIConnection-method}
\title{Create a simple table.}
\usage{
sqlCreateTable(con, name, fields, temporary = FALSE, ...)

\S4method{sqlCreateTable}{DBIConnection}(con, name, fields, temporary = FALSE,
  ...)
}
\arguments{
\item{con}{A database connection.}

\item{name}{Name of the table. Escaped with
\code{\link[DBI]{dbQuoteIdentifier}}.}

\item{fields}{A named character vector. Names are column names, values
are types. Names are escaped with \code{\link[DBI]{dbQuoteIdentifier}}.
Field types are unescaped.}

\item{temporary}{If \code{TRUE}, will generate a temporary table statement.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Exposes interface to simple \code{CREATE TABLE} commands. The default
method is ANSI SQL 99 compliant.
}
\examples{
sqlCreateTable(ANSI(), "my-table", c(a = "integer", b = "text"))
}

