% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DropTable.R
\docType{methods}
\name{sqlDropTable}
\alias{sqlDropTable}
\alias{sqlDropTable,DBIConnection-method}
\title{Drop table.}
\usage{
sqlDropTable(con, table, ...)

\S4method{sqlDropTable}{DBIConnection}(con, table, ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with
\code{\link[DBI]{dbQuoteIdentifier}}.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Drop table.
}
\examples{
sqlDropTable(ANSI(), "mtcars")
sqlDropTable(ANSI(), "mt\\nca\\"rs")
}

