% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/InsertInto.R
\docType{methods}
\name{sqlInsertInto}
\alias{sqlInsertInto}
\alias{sqlInsertInto,DBIConnection-method}
\title{Insert rows into a table.}
\usage{
sqlInsertInto(con, table, values, ...)

\S4method{sqlInsertInto}{DBIConnection}(con, table, values, ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with
\code{\link[DBI]{dbQuoteIdentifier}}.}

\item{values}{A data frame. Factors will be converted to character vectors.
Character vectors will be escaped with \code{\link[DBI]{dbQuoteString}}.}

\item{...}{Other arguments used by individual methods.}
}
\value{
A character vector with one element for each row in \code{values}.
}
\description{
Insert rows into a table.
}
\examples{
sqlInsertInto(ANSI(), "mtcars", head(mtcars))
sqlInsertInto(ANSI(), "iris", head(iris))
}

