% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TableDrop.R
\docType{methods}
\name{sqlTableDrop}
\alias{sqlTableDrop}
\alias{sqlTableDrop,DBIConnection-method}
\title{Drop table.}
\usage{
sqlTableDrop(con, table, ...)

\S4method{sqlTableDrop}{DBIConnection}(con, table, ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with
\code{\link[DBI]{dbQuoteIdentifier}}.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Drop table.
}
\examples{
sqlTableDrop(ANSI(), "mtcars")
sqlTableDrop(ANSI(), "mt\\nca\\"rs")
}

