% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TableInsertInto.R
\docType{methods}
\name{sqlTableInsertInto}
\alias{sqlTableInsertInto}
\alias{sqlTableInsertInto,DBIConnection-method}
\title{Insert rows into a table.}
\usage{
sqlTableInsertInto(con, table, values, ...)

\S4method{sqlTableInsertInto}{DBIConnection}(con, table, values, ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with
\code{\link[DBI]{dbQuoteIdentifier}}.}

\item{values}{A data frame. Factors will be converted to character vectors.
Character vectors will be escaped with \code{\link[DBI]{dbQuoteString}}.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Insert rows into a table.
}
\examples{
sqlTableInsertInto(ANSI(), "mtcars", head(mtcars))
sqlTableInsertInto(ANSI(), "iris", head(iris))
}

