% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TableInsertInto.R
\name{sqlTableInsertIntoTemplate}
\alias{sqlTableInsertIntoTemplate}
\title{Generated parameterised template for inserting rows.}
\usage{
sqlTableInsertIntoTemplate(con, table, values, row.names = NA, prefix = "?",
  ...)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Name of the table. Escaped with
\code{\link[DBI]{dbQuoteIdentifier}}.}

\item{values}{A data frame. Used only for the column names.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

  If \code{TRUE}, always translate row names to a column called "row_names".
  If \code{FALSE}, never translate row names. If \code{NA}, translate
  rownames only if they're a character vector.

  A string is equivalent to \code{TRUE}, but allows you to override the
  default name.

  For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{prefix}{Parameter prefix to put in front of column id.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Generated parameterised template for inserting rows.
}
\examples{
sqlTableInsertIntoTemplate(ANSI(), "iris", iris)

sqlTableInsertIntoTemplate(ANSI(), "mtcars", mtcars)
sqlTableInsertIntoTemplate(ANSI(), "mtcars", mtcars, row.names = FALSE)
}

