% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Transactions.R
\docType{methods}
\name{transactions}
\alias{sqlTransactionCommit}
\alias{sqlTransactionCommit,DBIConnection-method}
\alias{sqlTransactionRollback}
\alias{sqlTransactionRollback,DBIConnection-method}
\alias{sqlTransactionStart}
\alias{sqlTransactionStart,DBIConnection-method}
\alias{transactions}
\title{Start, commit or rollback a transaction}
\usage{
sqlTransactionStart(con, ...)

\S4method{sqlTransactionStart}{DBIConnection}(con, ...)

sqlTransactionCommit(con, ...)

\S4method{sqlTransactionCommit}{DBIConnection}(con, ...)

sqlTransactionRollback(con, ...)

\S4method{sqlTransactionRollback}{DBIConnection}(con, ...)
}
\arguments{
\item{con}{A database connection.}

\item{...}{Other arguments used by individual methods.}
}
\description{
Start, commit or rollback a transaction
}
\examples{
sqlTransactionStart(ANSI())
sqlTransactionCommit(ANSI())
sqlTransactionRollback(ANSI())
}

