% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQL}
\alias{SQL}
\title{Statistical Quantile Learning}
\usage{
SQL(
  data,
  q = 1,
  d = 4,
  lambda = 0.1,
  tol = 1e-04,
  max_iter = 30,
  max_cycles = 20,
  use_Rfast = FALSE,
  pretrain_tol = 5e-04,
  Sigma = NULL
)
}
\arguments{
\item{data}{A  real data matrix.}

\item{q}{The number of factors to be estimated.}

\item{d}{The number of B-splines basis functions.}

\item{lambda}{Tuning constant for penalyzation.}

\item{tol}{Tolerance for the qap optimization problem.}

\item{max_iter}{Maximum number of iterations for the qap algorithm.}

\item{max_cycles}{Maximum number for the cycles for the backfitting algorithm.}

\item{use_Rfast}{If TRUE, use package Rfast to compute the cross-product of the data matrix. This option may lead to speed gains if the number of variables is large.}

\item{pretrain_tol}{Tolerance for pretraining.}

\item{Sigma}{Cross-product of the centered data.}
}
\value{
A list with class \code{sql}.
\itemize{
\item \code{factor} a n times q matrix  containing the estimated latent factors.
\item \code{data} the data matrix.
\item \code{opt} a list containing the solution of the Quadratic Assignment Problem, the mean squared error for each iteration, and the convergence status.
\item \code{lambda} the tuning parameter.
\item \code{d} dimension of the basis functions.
\item \code{q} number of latent factors.
\item \code{intercepts} parameter vector of intercepts of length p.
\item \code{EV} Explained variance.
}
}
\description{
Estimate the Additive Factor Model with the Statistical Quantile Learning method.
The implementation makes use of penalyzed B-splines basis functions.
}
\examples{
set.seed(123456)
sim <- simulate_afm(n = 150, p = 200)
sql <- SQL(sim$data )
sql
plot(sql)
abs( cor(sim$factor, sql$factor) )
#==============
# q= 3 factor:
#==============
q <- 3
sim <- simulate_afm(n = 150, p = 200, q = q)
sql <- SQL(sim$data, q = q)
sql
abs( cor(sim$factor_z, sql$factor) )
}
